requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.3
playwright==1.45.0
httpx==0.27.0
websockets==11.0.3
pyyaml==6.0.1
python-dotenv==1.0.0
loguru==0.7.2
